Refactored to classes
